+++ elm-types-webpack-plugin.js
@@ -1,0 +1,29 @@
+const path = require('path');
+const convert = require('./build-helpers/generate-elm-ts-ipc.js');
+
+const elmModuleName = 'TsElmInterfaces';
+
+module.exports = class ElmTypesPlugin {
+  constructor(options) {
+    this.inputFile = options.inputFile;
+    this.outputPath = options.outputPath;
+  }
+
+  get outputFile() {
+    return path.join(this.outputPath, `${elmModuleName}.elm`);
+  }
+
+  apply(compiler) {
+    compiler.plugin('run', (compiler, callback) => {
+      console.log('Generate Elm type from typescript interfaces...');
+      convert(this.inputFile, this.outputFile, (err) => {
+        if (err) {
+          console.error('Error while generating Elm types', err);
+        } else {
+          console.log('Generation successful!');
+          callback();
+        }
+      });
+    });
+  }
+};
